@@ -12,15 +12,19 @@
   },
   "dependencies": {
     "@vue/cli-plugin-babel": "4.1.1",
-    "vue": "^2.6.11"
+    "vue": "^3.1.0",
+    "@vue/compat": "^3.1.0"
   },
   "devDependencies": {
     "@vue/cli-plugin-eslint": "4.1.1",
-    "@vue/cli-service": "4.1.1",
+    "@vue/cli-service": "^4.5.15",
+    "@vue/compiler-sfc": "^3.1.0",
     "babel-eslint": "^10.0.3",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.0.1",
-    "vue-template-compiler": "^2.6.11"
+    "node-sass": "^7.0.1",
+    "sass": "^1.45.1",
+    "sass-loader": "^10.1.1"
   },
   "eslintConfig": {
     "root": true,
