@@ -1,9 +1,5 @@
-import "./scss/main.css";
-
-import Vue from "vue";
+import { createApp } from "vue";
 import App from "./App.vue";
+import "./styles/media.scss";
 
-Vue.config.productionTip = false;
-new Vue({
-  render: (h) => h(App),
-}).$mount("#app");
+createApp(App).mount("#app");
