@@ -44,21 +44,6 @@
         </button>
 
         <button
-          v-if="playerID"
-          v-on:click="openMenu('playerMenu')"
-          v-on:keyup.enter="openMenu('playerMenu')"
-          class="player-button"
-          :class="playerMenuActive ? 'active' : ''"
-        >
-          <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24">
-            <path fill="none" d="M0 0h24v24H0V0z" />
-            <path
-              d="M9 13.75c-2.34 0-7 1.17-7 3.5V19h14v-1.75c0-2.33-4.66-3.5-7-3.5zM4.34 17c.84-.58 2.87-1.25 4.66-1.25s3.82.67 4.66 1.25H4.34zM9 12c1.93 0 3.5-1.57 3.5-3.5S10.93 5 9 5 5.5 6.57 5.5 8.5 7.07 12 9 12zm0-5c.83 0 1.5.67 1.5 1.5S9.83 10 9 10s-1.5-.67-1.5-1.5S8.17 7 9 7zm7.04 6.81c1.16.84 1.96 1.96 1.96 3.44V19h4v-1.75c0-2.02-3.5-3.17-5.96-3.44zM15 12c1.93 0 3.5-1.57 3.5-3.5S16.93 5 15 5c-.54 0-1.04.13-1.5.35.63.89 1 1.98 1 3.15s-.37 2.26-1 3.15c.46.22.96.35 1.5.35z"
-            />
-          </svg>
-        </button>
-
-        <button
           v-if="pitchMenu"
           v-on:click="openMenu('pitchMenu')"
           v-on:keyup.enter="openMenu('pitchMenu')"
@@ -69,6 +54,21 @@
             <path fill="none" d="M0 0h24v24H0z" />
             <path
               d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zM5.61 16.78C4.6 15.45 4 13.8 4 12s.6-3.45 1.61-4.78a5.975 5.975 0 0 1 0 9.56zM12 20c-1.89 0-3.63-.66-5-1.76 1.83-1.47 3-3.71 3-6.24S8.83 7.23 7 5.76C8.37 4.66 10.11 4 12 4s3.63.66 5 1.76c-1.83 1.47-3 3.71-3 6.24s1.17 4.77 3 6.24A7.963 7.963 0 0 1 12 20zm6.39-3.22a5.975 5.975 0 0 1 0-9.56C19.4 8.55 20 10.2 20 12s-.6 3.45-1.61 4.78z"
+            />
+          </svg>
+        </button>
+
+        <button
+          v-if="playerID"
+          v-on:click="openMenu('playerMenu')"
+          v-on:keyup.enter="openMenu('playerMenu')"
+          class="player-button"
+          :class="playerMenuActive ? 'active' : ''"
+        >
+          <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24">
+            <path fill="none" d="M0 0h24v24H0V0z" />
+            <path
+              d="M9 13.75c-2.34 0-7 1.17-7 3.5V19h14v-1.75c0-2.33-4.66-3.5-7-3.5zM4.34 17c.84-.58 2.87-1.25 4.66-1.25s3.82.67 4.66 1.25H4.34zM9 12c1.93 0 3.5-1.57 3.5-3.5S10.93 5 9 5 5.5 6.57 5.5 8.5 7.07 12 9 12zm0-5c.83 0 1.5.67 1.5 1.5S9.83 10 9 10s-1.5-.67-1.5-1.5S8.17 7 9 7zm7.04 6.81c1.16.84 1.96 1.96 1.96 3.44V19h4v-1.75c0-2.02-3.5-3.17-5.96-3.44zM15 12c1.93 0 3.5-1.57 3.5-3.5S16.93 5 15 5c-.54 0-1.04.13-1.5.35.63.89 1 1.98 1 3.15s-.37 2.26-1 3.15c.46.22.96.35 1.5.35z"
             />
           </svg>
         </button>
@@ -314,6 +314,6 @@ export default {
 };
 </script>
 
-<style scoped>
-@import "../scss/menu.min.css";
+<style lang="scss">
+@import "../styles/menu.scss";
 </style>
