@@ -53,10 +53,7 @@
 </template>
 
 <script>
-import { selectableMixin } from "../utility/selectable";
-
 export default {
-  mixins: [selectableMixin],
   props: {
     pitches: {
       type: Array,
@@ -156,6 +153,6 @@ export default {
 };
 </script>
 
-<style scoped lang="css">
-@import "../scss/pitch-list.min.css";
+<style lang="scss">
+@import "../styles/pitch-list.scss";
 </style>
